fix some architecture in models

+++ src/main/java/back/SportApp/Exercise/ExerciseService.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface ExerciseService {
     public Exercise create(Exercise exercise);
     public List<Exercise> findAll();
-    public Exercise findById(Long id);
-    public void deleteById(Long id);
+    public Exercise findById(Integer id);
+    public void deleteById(Integer id);
     public void update(Exercise exercise);
 }
